Pass acronyms in during call
Having the acronyms as a global variable meant that it was very
difficult to have different acronyms for different hash values, and this
caused us some problems when trying to use Awrence with Sidekiq (as
different jobs were modifying the global acronyms).

Passing them in during the call means that these acronyms are only
defined once they are necessary. Where these are stored is now a concern
of the user, not Awrence.